refactor: 💡 correct naming convention
