+++ src/items/applications/usecases/index.ts
@@ -1,4 +1,5 @@
 import { Provider } from '@nestjs/common';
-import { UpdateItemUsecase } from './update-item.usecase';
+import { UpdateItemUsecase } from './updateItem.usecase';
+import { GetItemsUseCase } from './getItems.usecase';
 
-export const UseCases: Provider[] = [UpdateItemUsecase];
+export const UseCases: Provider[] = [UpdateItemUsecase, GetItemsUseCase];
